Separate connection and endpoint logic
