+++ pymarsys/__init__.py
@@ -1,2 +1,3 @@
 __version__ = 0.1
+from .connections import AsyncConnection, SyncConnection
 from .emarsys import Emarsys
